refactor mock function
